Remove NumPy import from various core files
